@Henry Li
don't fork the repository since you are a contributor already; just create a new branch if you want to propose changes. Later, we can merge the two branches and accept them if we need them, or just delete the branch. This is a lot faster than making a whole other repository.