Updated camera code
Modified cam_sort so that it is now based off of percentages and not
pixels for easier usage. Went through code and cleaned up multiple
things such as superfluous function calls, and added comments that were
previously in the LinkCameraSorting project but were not transferred
over. Finally, added several print statements in the camera code to give
warnings and make testing easier.